backend-django/app/requirements_full.txt
asgiref>=3.4.1,<3.5
beautifulsoup4>=4.10.0,<4.11
bleach>=5.0.1,<6.0
celery>=5.2.3,<5.3
channels-redis>=3.3.1,<3.4
channels>=3.0.4,<3.1
coreapi>=2.3.3,<2.4
cryptography>=2.9.2,<2.10
uvicorn[standard]>=0.17.6,<0.18
gunicorn>=20.1.0,<20.2
defusedxml>=0.7.1,<0.8
django-admin-list-filter-dropdown>=1.0.3,<1.1
django-admin-sortable2>=1.0.4,<1.1
django-admin-autocomplete-filter>=0.7.1,<0.8
django-admin-rangefilter>=0.8.4,<0.9
django-anexia-monitoring>=1.3.0,<1.4
django-auth-ldap>=2.1.1,<2.2
django-ckeditor>=6.1.0,<6.2
django-cleanhtmlfield>=1.2.0,<1.3
django-cors-headers>=3.10.1,<3.11
django-dbbackup>=3.3.0,<3.4
django-filter>=21.1,<23.1
django-ipware>=3.0.7,<3.1
django-json-widget>=1.1.1,<1.2
django-memoize>=2.3.1,<2.4
django-redis-cache>=3.0.1,<3.1
django-request-cache>=1.2,<1.3
django-rest-multitokenauth>=1.5.0,<1.6
django-rest-passwordreset>=1.2.1,<1.3
django-userforeignkey>=0.5,<0.6
django-mptt>=0.13.4,<0.14
djangorestframework-xml>=2.0.0,<2.1
djangorestframework>=3.12.4,<3.13
drf-nested-routers>=0.93.4,<0.94
git+https://github.com/anx-bhagmann/djradicale
django-changeset>=1.1.0,<1.2
django-dbsettings>=1.1.0,<1.2
gitdb>=4.0.9,<4.1
# Pinned as 3.1.19 introduced an import error: ImportError: cannot import name
# TODO: Upgrade, 'TypeGuard' from 'typing_extensions'
GitPython>=3.1.27,<3.2
html2text>=2020.1.16,<2020.2
html5lib>=1.1,<1.2
lxml>=4.6.3,<4.7
Markdown>=3.4.1,<3.5
importlib-metadata>=4.12.0,<5.0.0
Pillow>=8.2.0,<8.3
# pinned for now: https://github.com/psycopg/psycopg2/issues/1293
psycopg2-binary>=2.8.6,<2.9
PyJWT>=1.7.1,<1.8
python-dateutil>=2.8.2,<2.9
redis>=3.5.3,<3.6
service_identity>=21.1.0,<21.2
time-machine>=2.4.1<2.5
vobject>=0.9.6.1,<0.10
WeasyPrint>=52.5,<53
zipstream>=1.1.4,<1.2
pyYAML>=5.3.1,<5.4
drf-yasg>=1.20.0,<1.21
pika>=1.2.0,<1.3
six>=1.16.0,<1.17
django-light
Unidecode>=1.3.4,<1.4
WsgiDAV==4.0.*
